Refactor the Makefile for OS X
Now there are "toolchain" commands like install-python-and-packages that delegate to "real" commands like install-python and install-python-packages, in the correct order.

Those "real" commands only delegate to real dependencies, not to commands that they think you might be interested in.

This makes it easier to install full toolchains ("I need Python and the usual packages on this computer" -> run install-python-and-packages) and debug failures ("Couldn't uninstall Python packages, let me see why" -> look at dependencies for uninstall-python-packages).